requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
tensorflow==2.15.0
numpy>=1.24.3
python-multipart==0.0.9
pillow==10.2.0
python-dotenv==1.0.1
cors==1.0.1
tensorflowjs==4.13.0
h5py==3.10.0 